# app/swagger.yaml
swagger: "2.0"
info:
  title: Medical Records API
  description: API untuk mengelola data rekam medis pasien, pengguna, laporan, dan dokumen medis lainnya.
  version: 1.0.0
host: localhost:5000
basePath: /
schemes:
  - http
produces:
  - application/json
consumes:
  - application/json
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: "Masukkan token JWT dengan format: 'Bearer {token}'"
security:
  - bearerAuth: []
...
definitions:
...
  CPPT:
    type: object
    properties:
      id:
        type: integer
      patient_id:
        type: integer
      tanggal:
        type: string
        format: date-time
      user_id:
        type: integer
      subjective:
        type: string
      objective:
        type: string
      assessment:
        type: string
      plan:
        type: string
      keterangan:
        type: string
      dokter:
        type: string
      signature:
        type: string
...
  Query:
    type: object
    properties:
      query:
        type: string
        description: "Teks atau kueri untuk diproses oleh AI atau pencarian."
paths:
...
  /cppt/:
    get:
      tags:
        - CPPT
      summary: Mendapatkan semua data CPPT
      security: []
      responses:
        '200':
          description: Sukses
    post:
      tags:
        - CPPT
      summary: Membuat entri CPPT baru menggunakan AI
      security: [] # Endpoint ini tidak memiliki dekorator JWT
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              query:
                type: string
              patient_id:
                type: integer
              perawat_id:
                type: integer
      responses:
        '201':
          description: CPPT berhasil dibuat
  /cppt/search:
    post:
      tags:
        - CPPT
      summary: Mencari data CPPT
      security: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/Query'
      responses:
        '200':
          description: Hasil pencarian
  /cppt/{cppt_id}:
    parameters:
      - name: cppt_id
        in: path
        required: true
        type: integer
    get:
      tags:
        - CPPT
      summary: Mendapatkan detail CPPT
      security: []
      responses:
        '200':
          description: Sukses
    put:
      tags:
        - CPPT
      summary: Memperbarui CPPT
      security: []
      parameters:
        - in: body
          name: body
          schema:
            $ref: '#/definitions/CPPT'
      responses:
        '200':
          description: CPPT berhasil diperbarui
    delete:
      tags:
        - CPPT
      summary: Menghapus CPPT
      security: []
      responses:
        '200':
          description: CPPT berhasil dihapus
...
  /intervensi/:
    get:
      tags:
        - Intervensi
      summary: Mendapatkan semua intervensi
      responses:
        '200':
          description: Daftar intervensi
          schema:
            type: array
            items:
              $ref: '#/definitions/Intervensi'

    post:
      tags:
        - Intervensi
      summary: Membuat intervensi baru dengan bantuan AI & referensi SIKI
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - query
              - patient_id
              - user_id
            properties:
              query:
                type: string
              patient_id:
                type: integer
              user_id:
                type: integer
      responses:
        '201':
          description: Intervensi berhasil dibuat
  /intervensi/{intervensi_id}:
    parameters:
      - name: intervensi_id
        in: path
        required: true
        type: integer
    get:
      tags:
        - Intervensi
      summary: Mendapatkan detail intervensi berdasarkan ID
      responses:
        '200':
          description: Detail intervensi
          schema:
            $ref: '#/definitions/Intervensi'

    put:
      tags:
        - Intervensi
      summary: Memperbarui data intervensi berdasarkan ID
      parameters:
        - in: body
          name: body
          schema:
            $ref: '#/definitions/Intervensi'
      responses:
        '200':
          description: Intervensi berhasil diperbarui

    delete:
      tags:
        - Intervensi
      summary: Menghapus intervensi berdasarkan ID
      responses:
        '200':
          description: Intervensi berhasil dihapus
  /intervensi/search:
    post:
      tags:
        - Intervensi
      summary: Mencari intervensi berdasarkan teks implementasi atau evaluasi
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/Query'
      responses:
        '200':
          description: Hasil pencarian intervensi
          schema:
            type: array
            items:
              $ref: '#/definitions/Intervensi'